% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/digitscount.R
\name{digitscount}
\alias{digitscount}
\title{count the numbers of the first digits of numeric vector data}
\usage{
digitscount(v)
}
\arguments{
\item{v}{a numerical vector}
}
\value{
A length 9 vector with ith component represents the number of first digits to be i in the input vector v.
}
\description{
count the numbers of the first digits(0 excluded) of numeric vector data
}
\examples{
v<-c(0,0,0,1.2,265.7,345.6,78,89,-91,-90.9,0.1,-0.02)
digitscount(v)
}

